logic of dog spawner separated from App.js to prevent re-renders
