+++ src/App.js
@@ -1,5 +1,6 @@
 import Music from './components/Music/Music'
 import Dog from './components/Dog/Dog'
+import DogSpawner from './containers/DogSpawner/DogSpawner'
 import Gradient from './components/Gradient/Gradient'
 /*
 Import linear gradients API
@@ -12,46 +13,13 @@ import './App.css';
 
 class App extends Component {
 
-  constructor(props){
-    super(props);
-    this.state = {
-      list: [true]
-    };
-  }
-
-
-  componentDidMount() {
-    //a function to add dog to the end of the array
-    const addDog = () => {
-      console.log("bang",this.state.list);
-      this.setState((prevState) => ({
-        list: [...prevState.list, true]
-      }))
-    }
-    //
-    // const removeDog = () => {
-    //   console.log("remove dog?")
-    //   this.setState((prevState) => {list: prevState.list.shift()})
-    // }
-
-    //every 1000 add dog
-    setInterval(() => addDog(), 5000)
-    //wait for 10s, remove a dog every s
-    // setTimeout(() => setInterval(() => removeDog(), 1000), 10000)
-  }
-
   render() {
 
     return (
       <div>
       <Gradient>
-        {this.state.list.map((data, index) => {
-          return <Dog />
-        })}
-
+      <DogSpawner />
       </Gradient>
-
-
       <Music />
       </div>
     );
